Move getContentConfig() to contentConfig.ts.
