+++ src/components/ChatBar.tsx
@@ -26,9 +26,9 @@ import {
   getLumosOptions,
 } from "../pages/Options";
 import { getHtmlContent } from "../scripts/content";
+import { getContentConfig } from "../contentConfig";
 import "@chatscope/chat-ui-kit-styles/dist/default/styles.min.css";
 import "./ChatBar.css";
-import { ContentConfig } from "../contentConfig";
 
 class LumosMessage {
   constructor(
@@ -79,38 +79,6 @@ const ChatBar: React.FC = () => {
 
     setChatContainerHeight(newChatContainerHeight);
     chrome.storage.local.set({ chatContainerHeight: newChatContainerHeight });
-  };
-
-  /**
-   * Return the content config that matches the current URL path.
-   *
-   * Each URL path can have a custom content config. For example,
-   * domain.com/path1 and domain.com/path2 can have different content
-   * configs. Additionally, content config paths can be nested. For
-   * example, domain.com/path1/subpath1 and domain.com/path1. In
-   * this case, the function will try to match the longest path first.
-   *
-   * Subdomains are also matched. If no matching path is found, null is
-   * returned.
-   */
-  const getContentConfig = (
-    url: URL,
-    contentConfig: ContentConfig,
-  ): null | {
-    chunkSize: number;
-    chunkOverlap: number;
-    selectors: string[];
-    selectorsAll: string[];
-  } => {
-    const searchPath = `${url.hostname}${url.pathname}`;
-
-    // Order keys (paths) of contentConfig in reverse order and check if any
-    // key is a substring of searchPath. This will find the longest matching
-    // key (path).
-    const paths = Object.keys(contentConfig).sort().reverse();
-    const matchingPath = paths.find((path) => searchPath.includes(path));
-
-    return matchingPath ? contentConfig[matchingPath] : null;
   };
 
   const promptWithContent = async () => {
